+++ packages/framework/src/hooks/index.ts
@@ -10,3 +10,5 @@ export * from "./useStyles";
 export * from "./useHtmlPassThrough";
 export * from "./useEnsembleUser";
 export * from "./useThemeContext";
+export * from "./useEventContext";
+export * from "./useLanguageScope";