+++ packages/framework/src/shared/common.ts
@@ -4,10 +4,45 @@ export type Expression<T> = string | T;
 
 export const isExpression = (
   maybeExpression: Expression<unknown>,
+): maybeExpression is string => {
+  if (!isString(maybeExpression)) {
+    return false;
+  }
+
+  const regex = /\$\{[^}]+\}/g;
+  return regex.test(maybeExpression);
+};
+
+export const isCompoundExpression = (input: string): boolean => {
+  const placeholders = [];
+  let stack = 0,
+    start = -1;
+
+  for (let i = 0; i < input.length; i++) {
+    if (input[i] === "$" && input[i + 1] === "{") {
+      if (stack === 0) start = i;
+      stack++;
+      i++;
+    } else if (input[i] === "}") {
+      stack--;
+      if (stack === 0 && start !== -1) {
+        placeholders.push(input.slice(start, i + 1));
+        start = -1;
+      }
+    }
+  }
+
+  return (
+    placeholders.length > 1 ||
+    (placeholders.length === 1 &&
+      input.replace(placeholders[0], "").trim().length > 0)
+  );
+};
+
+export const isTranslationKey = (
+  maybeExpression: Expression<unknown>,
 ): maybeExpression is string =>
-  isString(maybeExpression) &&
-  maybeExpression.startsWith("${") &&
-  maybeExpression.endsWith("}");
+  isString(maybeExpression) && maybeExpression.startsWith("r@");
 
 export const sanitizeJs = (string: string): string => {
   const trimmedString = string.trim();
@@ -29,9 +64,31 @@ export const findExpressions = (
   Object.entries(obj).forEach(([key, value]) => {
     const curPath = path.concat(key);
     if (isExpression(value)) {
-      expressionMap.push([curPath.join("."), value]);
+      expressionMap.push([
+        curPath.join("."),
+        isCompoundExpression(value) ? `\${\`${value || ""}\`}` : value,
+      ]);
     } else if (isObject(value)) {
       findExpressions(value, curPath, expressionMap);
+    }
+  });
+};
+
+export const findTranslationKeys = (
+  obj?: object,
+  path: string[] = [],
+  translationMap: string[][] = [],
+): void => {
+  if (!obj) {
+    return;
+  }
+
+  Object.entries(obj).forEach(([key, value]) => {
+    const curPath = path.concat(key);
+    if (isTranslationKey(value)) {
+      translationMap.push([curPath.join("."), value.replace("r@", "")]);
+    } else if (isObject(value)) {
+      findTranslationKeys(value, curPath, translationMap);
     }
   });
 };