+++ packages/runtime/src/runtime/hooks/useNavigateModal.tsx
@@ -7,25 +7,24 @@ import {
   useEnsembleStorage,
   useScreenContext,
 } from "@ensembleui/react-framework";
-import { cloneDeep, isObject, isString, merge, set } from "lodash-es";
+import { cloneDeep, isString, set } from "lodash-es";
 import { useCallback, useContext, useMemo } from "react";
-// FIXME: refactor
-// eslint-disable-next-line import/no-cycle
-import { EnsembleScreen } from "../screen";
 import { ModalContext } from "../modal";
 import { EnsembleRuntime } from "../runtime";
+// FIXME: refactor
+// eslint-disable-next-line import/no-cycle
+import { navigateModalScreen } from "../navigation";
 import type { EnsembleActionHook } from "./useEnsembleAction";
 
 export const useNavigateModalScreen: EnsembleActionHook<
   NavigateModalScreenAction
 > = (action?: NavigateModalScreenAction) => {
-  const { openModal } = useContext(ModalContext) || {};
+  const modalContext = useContext(ModalContext);
   const screenContext = useScreenContext();
   const customScope = useCustomScope();
   const storage = useEnsembleStorage();
 
   const isStringAction = isString(action);
-  const screenName = isStringAction ? action : action?.name;
 
   const title = useMemo(() => {
     if (!isStringAction && action?.title && !isString(action.title)) {
@@ -33,56 +32,31 @@ export const useNavigateModalScreen: EnsembleActionHook<
     }
   }, [isStringAction, action]);
 
-  const { matchingScreen } = useMemo(() => {
-    const screen = screenContext?.app?.screens.find(
-      (s) => s.name.toLowerCase() === screenName?.toLowerCase(),
-    );
-    return { matchingScreen: screen };
-  }, [screenContext, screenName]);
+  const callback = useCallback(
+    (args: unknown) => {
+      if (!action || !screenContext || !modalContext) {
+        return;
+      }
 
-  const callback = useCallback(() => {
-    if (!matchingScreen) {
-      return;
-    }
-
-    const inputs =
-      !isString(action) && action?.inputs ? cloneDeep(action.inputs) : {};
-    if (screenContext) {
-      const expressionMap: string[][] = [];
-      findExpressions(inputs, [], expressionMap);
-      expressionMap.forEach(([path, value]) => {
-        const result = evaluate(screenContext, value, {
-          ensemble: { storage },
-          ...customScope,
+      const inputs =
+        !isString(action) && action.inputs ? cloneDeep(action.inputs) : {};
+      if (screenContext) {
+        const expressionMap: string[][] = [];
+        findExpressions(inputs, [], expressionMap);
+        expressionMap.forEach(([path, value]) => {
+          const result = evaluate(screenContext, value, {
+            ensemble: { storage },
+            ...customScope,
+            ...(args as { [key: string]: unknown }),
+          });
+          set(inputs, path, result);
         });
-        set(inputs, path, result);
-      });
-    }
+      }
 
-    // modal options
-    const modalOptions = {
-      maskClosable: true,
-      hideCloseIcon: true,
-      hideFullScreenIcon: true,
-      padding: "12px",
-    };
-    if (isObject(action)) {
-      merge(modalOptions, action, { ...action.styles }, { title });
-    }
-
-    openModal?.(
-      <EnsembleScreen inputs={inputs} screen={matchingScreen} />,
-      modalOptions,
-    );
-  }, [
-    matchingScreen,
-    action,
-    screenContext,
-    openModal,
-    storage,
-    customScope,
-    title,
-  ]);
+      navigateModalScreen(action, screenContext, modalContext, inputs, title);
+    },
+    [action, screenContext, modalContext, title, storage, customScope],
+  );
 
   return { callback };
 };